moved to HiveTableUtil
